Add parameters to force specific IP versions
In some cases IPv4 seems to yield better results than IPv6. So give the user
paramters to force a specific protocol version.

In order to restrict curl specific defines and functions to fetch_data_string
own flags have been introduced, which are later on mapped to CURL_IPRESOLVE_V4
and CURL_IPRESOLVE_V6.

CAVEAT: It is possible to force IPv6, but this will currently fail at runtime as
status.raumzeitlabor.de does not resolve to IPv6.