Avoid hitting the cosm rate limit
The current implementation of ncrzlstat pulled data from cosm every time a key
was pressed or the window was resized. It was not unusual to hit cosms rate
limit after three minutes.

This commit changes the implementation in a way, that only after a given refresh
timeout the data is pulled. The display will still be repainted after every
keypress or SIGWINCH received. This is far from perfect, but it is good enough
at the moment.

Currently, I am working on a non-blocking display, that only repaints itself if
new data was pulled or a SIGWINCH was received.