Workaround issue in older ncurses implementations
A bug in GNU ncurses 5.7, and probably other versions, prevents ncurses default
SIGWINCH handler from being triggered. You can work around this issue by calling
doupdate() just before the call to getch(). At least with the latest version of
GNU ncurses 5.9 this issue seems to be fixed.

So, if the ui is not updated after resizing the terminal, try to build with
OLDCURSES defined.

This is the default, if you are using the BSDmakefile to build ncrzlstat.